requirements.txt
django
djangorestframework-simplejwt
django-extensions
drf_spectacular
PyJWT
djangorestframework
python-dotenv